fix: refactor dependency injection
