+++ src/Interfaces/IContext.cs
@@ -5,4 +5,5 @@ public interface IContext
   string? ApiKey { get; }
   int AppId { get; }
   string OutputDirectory { get; }
+  LogEventLevel LogLevel { get; }
 }
