+++ src/Models/Context.cs
@@ -5,11 +5,13 @@ public class Context : IContext
   public string? ApiKey { get; }
   public int AppId { get; }
   public string OutputDirectory { get; }
+  public LogEventLevel LogLevel { get; }
 
-  public Context(string? apiKey, int appId, string outputDirectory)
+  public Context(string? apiKey, int appId, string outputDirectory, LogEventLevel logLevel)
   {
     ApiKey = apiKey;
     AppId = appId;
     OutputDirectory = outputDirectory;
+    LogLevel = logLevel;
   }
 }
