+++ src/OnspringAttachmentReporter.csproj
@@ -15,7 +15,6 @@
     <PackageReference Include="Microsoft.Extensions.Configuration" Version="8.0.0-preview.1.23110.8" />
     <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="8.0.0-preview.1.23110.8" />
     <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="8.0.0-preview.1.23110.8" />
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="7.0.1" />
     <PackageReference Include="Serilog" Version="3.0.0-dev-01862" />
     <PackageReference Include="Serilog.Formatting.Compact" Version="1.1.1-dev-00944" />
     <PackageReference Include="Serilog.Sinks.Console" Version="4.1.1-dev-00907" />
