+++ tests/Usings.cs
@@ -11,5 +11,7 @@ global using OnspringAttachmentReporter.Models;
 global using OnspringAttachmentReporter.Services;
 
 global using Serilog;
+global using Serilog.Core;
+global using Serilog.Events;
 
 global using Xunit;
